fix: make page files thin entry points, move all logic to modules components
Pages are now one-liner wrappers (SettingsSubdomainPage, SettingsCustomDomainPage)
that just render the modules components (SettingsSubdomain, SettingsCustomDomain).
All logic, layout, save buttons, and modals live in the modules components
with no props drilling.

Made-with: Cursor
