+++ packages/twenty-front/src/modules/settings/domains/components/SettingsCustomDomain.tsx
@@ -1,17 +1,24 @@
 /* @license Enterprise */
 import { currentWorkspaceState } from '@/auth/states/currentWorkspaceState';
+import { SaveAndCancelButtons } from '@/settings/components/SaveAndCancelButtons/SaveAndCancelButtons';
+import { SettingsPageContainer } from '@/settings/components/SettingsPageContainer';
+import { CheckCustomDomainValidRecordsEffect } from '@/settings/domains/components/CheckCustomDomainValidRecordsEffect';
+import { SettingsDomainRecords } from '@/settings/domains/components/SettingsDomainRecords';
+import { useSettingsCustomDomain } from '@/settings/domains/hooks/useSettingsCustomDomain';
+import { customDomainRecordsState } from '@/settings/domains/states/customDomainRecordsState';
 import { TextInput } from '@/ui/input/components/TextInput';
-import { styled } from '@linaria/react';
-import { useLingui } from '@lingui/react/macro';
+import { SubMenuTopBarContainer } from '@/ui/layout/page/components/SubMenuTopBarContainer';
 import { useAtomStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomStateValue';
+import { Trans, useLingui } from '@lingui/react/macro';
+import { styled } from '@linaria/react';
+import { SettingsPath } from 'twenty-shared/types';
+import { getSettingsPath } from 'twenty-shared/utils';
 import { H2Title, IconReload, IconTrash } from 'twenty-ui/display';
 import { Button, ButtonGroup } from 'twenty-ui/input';
 import { Section } from 'twenty-ui/layout';
 import { themeCssVariables } from 'twenty-ui/theme-constants';
-import { SettingsDomainRecords } from '@/settings/domains/components/SettingsDomainRecords';
-import { CheckCustomDomainValidRecordsEffect } from '@/settings/domains/components/CheckCustomDomainValidRecordsEffect';
+import { useNavigateSettings } from '~/hooks/useNavigateSettings';
 import { useCheckCustomDomainValidRecords } from '@/settings/domains/hooks/useCheckCustomDomainValidRecords';
-import { customDomainRecordsState } from '@/settings/domains/states/customDomainRecordsState';
 
 const StyledDomainFormWrapper = styled.div`
   display: flex;
@@ -36,72 +43,96 @@ const StyledRecordsWrapper = styled.div`
   }
 `;
 
-type SettingsCustomDomainProps = {
-  value: string | null;
-  onChange: (value: string) => void;
-  onDelete: () => void;
-  error?: string;
-};
-
-export const SettingsCustomDomain = ({
-  value,
-  onChange,
-  onDelete,
-  error,
-}: SettingsCustomDomainProps) => {
-  const { customDomainRecords, isLoading } = useAtomStateValue(
-    customDomainRecordsState,
-  );
-
+export const SettingsCustomDomain = () => {
+  const navigate = useNavigateSettings();
+  const { t } = useLingui();
+  const currentWorkspace = useAtomStateValue(currentWorkspaceState);
+  const { customDomainRecords, isLoading: isRecordsLoading } =
+    useAtomStateValue(customDomainRecordsState);
   const { checkCustomDomainRecords } = useCheckCustomDomainValidRecords();
 
-  const currentWorkspace = useAtomStateValue(currentWorkspaceState);
-
-  const { t } = useLingui();
+  const {
+    customDomain,
+    error,
+    isSubmitting,
+    isSaveDisabled,
+    handleChange,
+    handleDelete,
+    handleSave,
+  } = useSettingsCustomDomain();
 
   return (
-    <Section>
-      <H2Title
-        title={t`Custom Domain`}
-        description={t`Set the name of your custom domain and configure your DNS records.`}
-      />
-      <CheckCustomDomainValidRecordsEffect />
-      <StyledDomainFormWrapper>
-        <TextInput
-          value={value ?? ''}
-          type="text"
-          onChange={onChange}
-          placeholder="crm.yourdomain.com"
-          error={error}
-          fullWidth
+    <SubMenuTopBarContainer
+      title={t`Custom Domain`}
+      links={[
+        {
+          children: <Trans>Workspace</Trans>,
+          href: getSettingsPath(SettingsPath.Workspace),
+        },
+        {
+          children: <Trans>Domains</Trans>,
+          href: getSettingsPath(SettingsPath.Domains),
+        },
+        { children: <Trans>Custom Domain</Trans> },
+      ]}
+      actionButton={
+        <SaveAndCancelButtons
+          onCancel={() => navigate(SettingsPath.Domains)}
+          isSaveDisabled={isSaveDisabled}
+          isLoading={isSubmitting}
+          onSave={handleSave}
         />
-        {currentWorkspace?.customDomain && (
-          <StyledButtonGroupContainer>
-            <ButtonGroup>
-              <StyledButtonContainer>
-                <Button
-                  isLoading={isLoading}
-                  Icon={IconReload}
-                  title={t`Reload`}
-                  variant="primary"
-                  onClick={checkCustomDomainRecords}
-                  type="button"
-                />
-              </StyledButtonContainer>
-              <StyledButtonContainer>
-                <Button Icon={IconTrash} variant="primary" onClick={onDelete} />
-              </StyledButtonContainer>
-            </ButtonGroup>
-          </StyledButtonGroupContainer>
-        )}
-      </StyledDomainFormWrapper>
-      {currentWorkspace?.customDomain && (
-        <StyledRecordsWrapper>
-          {customDomainRecords && (
-            <SettingsDomainRecords records={customDomainRecords.records} />
+      }
+    >
+      <SettingsPageContainer>
+        <Section>
+          <H2Title
+            title={t`Custom Domain`}
+            description={t`Set the name of your custom domain and configure your DNS records.`}
+          />
+          <CheckCustomDomainValidRecordsEffect />
+          <StyledDomainFormWrapper>
+            <TextInput
+              value={customDomain}
+              type="text"
+              onChange={handleChange}
+              placeholder="crm.yourdomain.com"
+              error={error}
+              fullWidth
+            />
+            {currentWorkspace?.customDomain && (
+              <StyledButtonGroupContainer>
+                <ButtonGroup>
+                  <StyledButtonContainer>
+                    <Button
+                      isLoading={isRecordsLoading}
+                      Icon={IconReload}
+                      title={t`Reload`}
+                      variant="primary"
+                      onClick={checkCustomDomainRecords}
+                      type="button"
+                    />
+                  </StyledButtonContainer>
+                  <StyledButtonContainer>
+                    <Button
+                      Icon={IconTrash}
+                      variant="primary"
+                      onClick={handleDelete}
+                    />
+                  </StyledButtonContainer>
+                </ButtonGroup>
+              </StyledButtonGroupContainer>
+            )}
+          </StyledDomainFormWrapper>
+          {currentWorkspace?.customDomain && (
+            <StyledRecordsWrapper>
+              {customDomainRecords && (
+                <SettingsDomainRecords records={customDomainRecords.records} />
+              )}
+            </StyledRecordsWrapper>
           )}
-        </StyledRecordsWrapper>
-      )}
-    </Section>
+        </Section>
+      </SettingsPageContainer>
+    </SubMenuTopBarContainer>
   );
 };
