+++ packages/twenty-front/src/pages/settings/domains/SettingsCustomDomainPage.tsx
@@ -1,0 +1,4 @@
+/* @license Enterprise */
+import { SettingsCustomDomain } from '@/settings/domains/components/SettingsCustomDomain';
+
+export const SettingsCustomDomainPage = () => <SettingsCustomDomain />;
