+++ packages/twenty-front/src/pages/settings/domains/SettingsSubdomainPage.tsx
@@ -1,0 +1,3 @@
+import { SettingsSubdomain } from '@/settings/domains/components/SettingsSubdomain';
+
+export const SettingsSubdomainPage = () => <SettingsSubdomain />;
